Add feature toggle to turn the delete-before-migrate functionality off whilst performance is being investigated

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsremandandsentencingapi/legacy/service/MigrationService.kt
@@ -1,6 +1,7 @@
 package uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.legacy.service
 
 import org.slf4j.LoggerFactory
+import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.jpa.entity.AppearanceChargeEntity
@@ -82,11 +83,16 @@ class MigrationService(
   private val recallRepository: RecallRepository,
   private val recallSentenceRepository: RecallSentenceRepository,
   private val appearanceChargeRepository: AppearanceChargeRepository,
+  @Value("\${delete.before.migrate:false}")
+  private val deleteBeforeMigrate: Boolean,
 ) {
 
   @Transactional
   fun create(migrationCreateCourtCases: MigrationCreateCourtCases): MigrationCreateCourtCasesResponse {
-    deletePrisonerData(migrationCreateCourtCases.prisonerId)
+    if (deleteBeforeMigrate) {
+      deletePrisonerData(migrationCreateCourtCases.prisonerId)
+    }
+
     val tracking = MigrationDataTracking(
       migrationCreateCourtCases.prisonerId,
       serviceUserService.getUsername(),
